test(hosting): move module load tests
