+++ JavaScriptRuntime/Engine/Engine.cs
@@ -24,7 +24,7 @@ public class Engine
         {
             // Validate caller provided a valid entry point delegate.
             // Note: the delegate's Method/Module/Assembly is used to discover the compiled module assembly.
-            ArgumentException.ThrowIfNullOrEmpty(nameof(scriptEntryPoint));
+            ArgumentNullException.ThrowIfNull(scriptEntryPoint);
 
             // Configure per-thread services and install the Node-like synchronization context.
             // This enables timers/microtasks and other async behavior to run deterministically on this thread.
